pybind/mgr/snap-schedule: refactor schedule into two modules
This separates the Schedule class from the SnapSchedClient class.
Motivation is cleaner code layout and simpler testing.

(cherry picked from commit aee0981b418c5a9077b967ff089c6434cd42f7e8)
